MAPR-22888 ERROR No such file or directory after querying data from any table via HiveCLI

+++ ql/src/java/org/apache/hadoop/hive/ql/Context.java
@@ -556,9 +556,11 @@ public class Context {
       try {
         Path p = entry.getValue();
         FileSystem fs = p.getFileSystem(conf);
-        LOG.debug("Deleting scratch dir: {}",  p);
-        fs.delete(p, true);
-        fs.cancelDeleteOnExit(p);
+        if(fs.exists(p)) {
+          LOG.debug("Deleting scratch dir: {}", p);
+          fs.delete(p, true);
+          fs.cancelDeleteOnExit(p);
+        }
       } catch (Exception e) {
         LOG.warn("Error Removing Scratch: "
             + StringUtils.stringifyException(e));
